Move resources to separate file
