+++ public/app/features/provisioning/RepositoryStatusPage.tsx
@@ -30,13 +30,12 @@ import { isNotFoundError } from '../alerting/unified/api/util';
 
 import { ConfigForm } from './ConfigForm';
 import { ExportToRepository } from './ExportToRepository';
+import { RepositoryResources } from './RepositoryResources';
 import { RepositorySyncStatus } from './RepositorySyncStatus';
 import {
   useListJobQuery,
   useGetRepositoryFilesQuery,
   Repository,
-  ResourceListItem,
-  useGetRepositoryResourcesQuery,
   useListRepositoryQuery,
   useDeleteRepositoryFilesWithPathMutation,
 } from './api';
@@ -104,7 +103,7 @@ export default function RepositoryStatusPage() {
                   ))}
                 </TabsBar>
                 <TabContent>
-                  {tab === TabSelection.Resources && <ResourcesView repo={data} />}
+                  {tab === TabSelection.Resources && <RepositoryResources repo={data} />}
                   {tab === TabSelection.Files && <FilesView repo={data} />}
                   {tab === TabSelection.Jobs && <JobsView repo={data} />}
                   {tab === TabSelection.Sync && <RepositorySyncStatus repo={data} />}
@@ -219,95 +218,6 @@ function FilesView({ repo }: RepoProps) {
         <FilterInput placeholder="Search" autoFocus={true} value={searchQuery} onChange={setSearchQuery} />
       </Stack>
       <InteractiveTable columns={columns} data={data} pageSize={25} getRowId={(f: FileDetails) => String(f.path)} />
-    </Stack>
-  );
-}
-
-type ResourceCell<T extends keyof ResourceListItem = keyof ResourceListItem> = CellProps<
-  ResourceListItem,
-  ResourceListItem[T]
->;
-
-function ResourcesView({ repo }: RepoProps) {
-  const name = repo.metadata?.name ?? '';
-  const query = useGetRepositoryResourcesQuery({ name });
-  const [searchQuery, setSearchQuery] = useState('');
-  const data = [...(query.data?.items ?? [])].filter((Resource) =>
-    Resource.path.toLowerCase().includes(searchQuery.toLowerCase())
-  );
-  const columns: Array<Column<ResourceListItem>> = useMemo(
-    () => [
-      {
-        id: 'title',
-        header: 'Title',
-        sortType: 'string',
-        cell: ({ row: { original } }: ResourceCell<'title'>) => {
-          const { resource, name, title } = original;
-          if (resource === 'dashboards') {
-            return <a href={`/d/${name}`}>{title}</a>;
-          }
-          if (resource === 'folders') {
-            return <a href={`/dashboards/f/${name}`}>{title}</a>;
-          }
-          return <span>{title}</span>;
-        },
-      },
-      {
-        id: 'path',
-        header: 'Path',
-        sortType: 'string',
-        cell: ({ row: { original } }: ResourceCell<'path'>) => {
-          const { resource, name, path } = original;
-          if (resource === 'dashboards') {
-            return <a href={`/d/${name}`}>{path}</a>;
-          }
-          return <span>{path}</span>;
-        },
-      },
-      {
-        id: 'hash',
-        header: 'Hash',
-        sortType: 'string',
-        cell: ({ row: { original } }: ResourceCell<'hash'>) => {
-          const { hash } = original;
-          return <span title={hash}>{hash.substring(0, 7)}</span>;
-        },
-      },
-      {
-        id: 'folder',
-        header: 'Folder',
-        sortType: 'string',
-        cell: ({ row: { original } }: ResourceCell<'title'>) => {
-          const { folder } = original;
-          if (folder?.length) {
-            return <a href={`/dashboards/f/${folder}`}>{folder}</a>;
-          }
-          return <span></span>;
-        },
-      },
-    ],
-    []
-  );
-
-  if (query.isLoading) {
-    return (
-      <Stack justifyContent={'center'} alignItems={'center'}>
-        <Spinner />
-      </Stack>
-    );
-  }
-
-  return (
-    <Stack grow={1} direction={'column'} gap={2}>
-      <Stack gap={2}>
-        <FilterInput placeholder="Search" autoFocus={true} value={searchQuery} onChange={setSearchQuery} />
-      </Stack>
-      <InteractiveTable
-        columns={columns}
-        data={data}
-        pageSize={25}
-        getRowId={(r: ResourceListItem) => String(r.path)}
-      />
     </Stack>
   );
 }
